Edit: rename sign-in, sign-up folder
