+++ front-end/src/app/screens/index.ts
@@ -2,5 +2,5 @@ export * from "./admin";
 export * from "./carts";
 export * from "./home";
 export * from "./products";
-export * from "./sign-in";
-export * from "./sign-up";
+export * from "./signin";
+export * from "./signup";
